incomplete, first functions
an incomplete version of the functions that will be needed. Present is a function to generate stimuli for the visual search task, and an idea for a function to generate a (balanced) random trial order according to weights.
To do: GUI functions, logging functions for participant response, data saving structures, package skeleton