GUI and config update
GUI now includes all of the options needed for the VST settings, with pre-filled boxes where needed based on a config. The config is updated based on what the user fills in.

RunVST() now uses the config when calling TrialOrder(), the next goal is to use the resulting trial information to generate stimuli.