instructions, saving, randomising
progress made on showing an instruction screen, and results are now saved.
progress made on using the generated trial information, seemingly one problem to solve: getting the stimuli numbers correctly recognised as a list of integers
Because this is in a partially-solved state, this commit is not functional.

Still needed:
- Correct transition from instruction screen to first trial
- exit screen
- "current settings" screen!
- blocks